# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Neighbor Weather Check-ins API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CheckinWeatherType = typeof CheckinWeatherType[keyof typeof CheckinWeatherType];


export const CheckinWeatherType = {
  sunny: 'sunny',
  cloudy: 'cloudy',
  rainy: 'rainy',
  stormy: 'stormy',
  windy: 'windy',
  snowy: 'snowy',
  foggy: 'foggy',
} as const;

export interface Checkin {
  id: number;
  weatherType: CheckinWeatherType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  locationName?: string | null;
  createdAt: string;
  agrees: number;
}

export type CheckinInputWeatherType = typeof CheckinInputWeatherType[keyof typeof CheckinInputWeatherType];


export const CheckinInputWeatherType = {
  sunny: 'sunny',
  cloudy: 'cloudy',
  rainy: 'rainy',
  stormy: 'stormy',
  windy: 'windy',
  snowy: 'snowy',
  foggy: 'foggy',
} as const;

export interface CheckinInput {
  weatherType: CheckinInputWeatherType;
  description?: string;
  locationName?: string;
}

export interface CheckinSummaryItem {
  weatherType: string;
  count: number;
}

export interface CheckinSummary {
  items: CheckinSummaryItem[];
  totalToday: number;
  /** @nullable */
  mostCommon: string | null;
}

export type ListCheckinsParams = {
limit?: number;
weatherType?: string;
};