# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Neighbor Weather Check-ins API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckinWeatherType } from './checkinWeatherType';

export interface Checkin {
  id: number;
  weatherType: CheckinWeatherType;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  locationName?: string | null;
  createdAt: string;
  agrees: number;
}