# lib/api-zod/src/generated/types/checkinInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Neighbor Weather Check-ins API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckinInputWeatherType } from './checkinInputWeatherType';

export interface CheckinInput {
  weatherType: CheckinInputWeatherType;
  description?: string;
  locationName?: string;
}