# lib/api-zod/src/generated/types/checkinSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Neighbor Weather Check-ins API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckinSummaryItem } from './checkinSummaryItem';

export interface CheckinSummary {
  items: CheckinSummaryItem[];
  totalToday: number;
  /** @nullable */
  mostCommon: string | null;
}