# lib/api-zod/src/generated/types/checkinSummaryItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Neighbor Weather Check-ins API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinSummaryItem {
  weatherType: string;
  count: number;
}